Create ToolBarController
Parts concerning the toolbar were extracted to the ToolBarController
and some parts of the mediator class were either commented out or
completely removed.
